Moved cost calculus to a different class
